+++ R/conditions.R
@@ -1,7 +1,7 @@
 msg_testfile_does_not_exist <- function(x){
   message(
-    sprintf('The file `%s` does not exist. ', x),
-    'You can create it with testthis::test_skeleton().'
+    sprintf("The file '%s' does not exist. ", x),
+    "You can create it with testthis::test_skeleton()."
   )
 }
 
@@ -10,6 +10,6 @@ msg_testfile_does_not_exist <- function(x){
 
 require_rstudio <- function(){
   if(!requireNamespace("rstudioapi", quietly = TRUE)){
-    stop('This function requires access the RStudio API', call. = FALSE)
+    stop("This function requires access the RStudio API", call. = FALSE)
   }
 }