+++ R/test_this.R
@@ -1,6 +1,6 @@
 #' Test this file
 #'
-#' Runs testthat tests in a single .R file. If the file currently open in the
+#' Runs testthat tests in a single `.R` file. If the file currently open in the
 #' RStudio editor is called `my_function.R`, `test_this()` calls
 #' `testthat::test_file()` on \file{tests/testthat/test_my_function.R}. If
 #' the filename of the currently open file with starts with `test_` it will
@@ -32,19 +32,4 @@ test_this <- function(...){
   }
 
   invisible()
-}
-
-
-
-
-#' \code{lest_this()} Deprecated. Please us `test_this()` instead.
-#' @export
-#' @rdname test_this
-lest_this <- function(...){
-  .Deprecated(
-    msg = paste(
-      "'test_this()' now reloads package by default. 'lest_this()' is no",
-      "longer required and will be dropped in future versions."
-    ))
-  test_this(...)
 }