+++ tests/testthat/testdata/test_pkg/tests/testthat/testthat_parse_cases.R
@@ -17,3 +17,15 @@ test_that(
 test_that(
   code = expect_true(FALSE),
   "%barfun% is ok")
+
+
+for (i in 1){
+  test_that(paste0("blubb", i), {
+    expect_true(1 == 1)
+  })
+
+  test_that("foo", {
+    expect_true(1 == 2)
+  })
+}
+